refactor: using action in SearchForm's Form
