+++ app/all-fallen/loading.jsx
@@ -1,4 +1,4 @@
-import BouncyLine from "../components/Spinners/BouncyLine/BouncyLine";
+import BouncyLine from "../components/Spinners/BouncyLine";
 
 export default function Loading() {
   return <BouncyLine />;
