+++ app/components/AuthPopup/index.jsx
@@ -6,7 +6,7 @@ import styles from "./style.module.scss";
 import { signIn } from "next-auth/react";
 import { useEffect, useState } from "react";
 import GenericInput from "../GenericInput/index";
-import SpinningCircle from "../Spinners/SpinningCircle/SpinningCircle";
+import SpinningCircle from "../Spinners/SpinningCircle";
 import {
   resetPasswordFields,
   loginFields,
