+++ app/components/InvitationBubble/index.jsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { useRouter } from "next/navigation";
 import LargeBubble from "../LargeBubble";
+import { useRouter } from "next/navigation";
 
 const InvitationBubble = () => {
   const router = useRouter();
